Add files via upload
Online Medical Management System, titled “Medical Treatment”, is a web-based project developed in Java with the aim of providing a platform for online health issue resolve and find a disease name and recommended doctors. It enables to gives online pharmacy, precautions and doctor's etc.
 
The site should have proper users’ home management and admin login. There must be list of doctors for a disease. There have a contact details and address of doctors should be available for service of patients. There should be a facility of making a complaint to the admin for any kind of medical error. The site has an online pharmacy and gives a precaution to the patients. Admin is supposed to take add or edit of source code and it should be capable of viewing source code.
 
This is a second-year project in Java. Being a web-based project, it provides a platform for find disease, online pharmacy, precautions and doctors. You can access the full project report, documentation, project files and Java source code of Online Medical Treatment from the download link/button in this post. If you have any queries, drop them in the comments.